# api/openapi.yaml
openapi: 3.1.0
info:
  title: RockServer API
  version: 0.6.0
  description: >-
    Contract for RockServer operational health checks, deterministic station search, and the
    Windows voice-command JSON and streaming boundaries. `POST /api/v1/voice/command` resolves
    an already-recognized command and `GET /api/v1/voice/stream` upgrades to the
    provider-neutral WebSocket protocol for raw PCM audio and incremental transcripts.
    PostgreSQL may use provenance-compatible semantic ranking; metadata fallback preserves
    the same schemas when embeddings are absent or unavailable. Operations marked
    `x-rockserver-status: planned` are contract-first device-control designs and are not yet
    registered by the runtime. The server-resolved `station.play_stream` dispatch, device voice
    sessions, and the two device-facing
    catalog paths (`GET /api/v1/device-control/catalog/stations` and
    `GET /api/v1/device-control/catalog/search`) are implemented behind the RockserverBearer
    device session.
servers:
  - url: https://rockplatform.win
  - url: http://localhost
paths:
...
  /api/v1/device-control/catalog/stations:
    get:
      operationId: browseDeviceCatalog
      summary: Browse one bounded device-facing catalog page of station metadata
      x-rockserver-status: implemented
      description: >-
        Device-facing catalog browse for paired native devices that render a station list
        locally (ESP32 display/controller roles). Authentication is the short-lived
        RockserverBearer device session from `POST /api/v1/auth/device-session`; the server
        derives user and device identity from it and never trusts identity from parameters.
        Station metadata is account-independent, so v1 requires only a valid non-revoked
        device session and no additional device-control scope. The cursor model matches the
        public catalog: the cursor is the stable station ID of the last item of the previous
        page, at most 512 ASCII letters, digits or hyphens, and `next_cursor` is null exactly
        when no further page may exist. At most 20 stations are returned per page. Unknown or
        oversized parameters are rejected deterministically with 400. Responses never contain
        `stream_url`: a device starts playback only through server-resolved
        `station.play_station` dispatch, never through a catalog URL.
      security:
        - RockserverBearer: []
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 20, default: 20 }
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
            pattern: "^[A-Za-z0-9-]+$"
      responses:
        "200":
          description: One bounded station-metadata page without stream URLs.
          headers:
            Cache-Control: { schema: { type: string, example: no-store } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceCatalogPage" }
        "400":
          description: An unknown, malformed, or oversized pagination parameter was supplied.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/AuthenticationRequired" }
        "403":
          description: The authenticated device is revoked or forbidden without disclosing another target.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429":
          description: The caller exceeded the bounded device-catalog rate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: The station catalog is temporarily unavailable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/device-control/catalog/search:
    get:
      operationId: searchDeviceCatalog
      summary: Run one bounded ranked station search for a device session
      x-rockserver-status: implemented
      description: >-
        Device-facing ranked search reusing the existing deterministic station-search service
        (hard filters, deterministic ranking, semantic-provider failures degrade to metadata
        search). Authentication is the same RockserverBearer device session as catalog browse;
        the server derives identity from it. `q` is at most 128 characters, at most 20 ranked
        stations are returned, and the result is exactly one page: ranked search is not
        cursorable in the existing service, so no cursor parameter or `next_cursor` field
        exists on this path (two-path decision, 2026-09-09). Rank order is the array order.
        Unknown or oversized parameters are rejected deterministically with 400. Responses
        never contain `stream_url`.
      security:
        - RockserverBearer: []
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
          description: Natural-language station query; whitespace-only values are invalid.
        - name: locale
          in: query
          required: false
          schema:
            type: string
            default: en-US
            pattern: "^[A-Za-z]{2,3}(?:-[A-Za-z0-9]{2,8})*$"
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 20, default: 20 }
      responses:
        "200":
          description: One ranked station-metadata page without stream URLs and without a cursor.
          headers:
            Cache-Control: { schema: { type: string, example: no-store } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceSearchPage" }
        "400":
          description: An unknown, malformed, oversized, or whitespace-only query parameter was supplied.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/AuthenticationRequired" }
        "403":
          description: The authenticated device is revoked or forbidden without disclosing another target.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429":
          description: The caller exceeded the bounded device-catalog rate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: An unexpected catalog or search failure occurred.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: The station catalog is temporarily unavailable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
...
  securitySchemes:
    RockserverBrowserCookie:
      type: apiKey
      in: cookie
      name: rockserver_browser
      description: >-
        Secure, HttpOnly, SameSite browser session cookie issued after a verified passkey ceremony.
        State-changing browser requests also require the X-CSRF-Token header and the trusted Caddy
        proxy proof; those transport checks are enforced outside the bearer scheme.
    RockCastBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >-
        Legacy deployment-configured Bearer accepted when supplied to `/api/v1/voice/*` routes.
        Normal `/api/v1/search` and `/api/v1/voice/*` product routes are anonymous; this credential is not
        part of the account/device pairing UX and must not be exposed there.
    RockserverBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >-
        Short-lived opaque native-session access token issued by the existing pairing/session flow.
        It is the sole identity accepted by planned device-control operations and is also accepted
        by current native account and voice routes. Roles/capabilities are not token scopes;
        device-control policy is computed and enforced server-side.
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Short-lived, server-revocable administrator session token. It is separate from native and browser account credentials.
  responses:
    AuthenticationRequired:
      description: A required Bearer credential was absent, malformed, or invalid.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: authentication_required
            message: A valid Bearer token is required.
            request_id: req_01AUTH401
            details: {}
  schemas:
...
    DeviceStationDto:
      type: object
      additionalProperties: false
      required: [id, name, homepage_url, tags, language, country_code, codec, bitrate_kbps, health]
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 128
          description: Stable RockServer station identifier usable in `station.play_station`.
        name:
          type: string
          minLength: 1
        homepage_url:
          type: [string, "null"]
          format: uri
        tags:
          type: array
          items: { type: string }
        language:
          type: [string, "null"]
          description: ISO 639 language code when known.
        country_code:
          type: [string, "null"]
          pattern: "^[A-Z]{2}$"
        codec:
          type: [string, "null"]
        bitrate_kbps:
          type: [integer, "null"]
          minimum: 1
        health:
          type: string
          enum: [healthy, degraded, unknown]
      description: >-
        Station metadata for device-facing catalog pages. It deliberately has no `stream_url`,
        no ranking fields, and no provider or persistence identifiers: devices select a
        station by stable ID and playback is resolved server-side through
        `station.play_station` dispatch.
    DeviceCatalogPage:
      type: object
      additionalProperties: false
      required: [request_id, stations, next_cursor]
      properties:
        request_id:
          type: string
          minLength: 1
        stations:
          type: array
          maxItems: 20
          items: { $ref: "#/components/schemas/DeviceStationDto" }
        next_cursor:
          oneOf:
            - type: string
              minLength: 1
              maxLength: 512
              pattern: "^[A-Za-z0-9-]+$"
            - type: "null"
          description: Stable ID of the last returned station, or null when no further page may exist.
      description: >-
        Exactly one bounded browse page. The cursor is the last stable station ID of the
        previous page; the next page starts strictly after it in stable-ID order.
    DeviceSearchPage:
      type: object
      additionalProperties: false
      required: [request_id, stations]
      properties:
        request_id:
          type: string
          minLength: 1
        stations:
          type: array
          maxItems: 20
          items: { $ref: "#/components/schemas/DeviceStationDto" }
      description: >-
        Exactly one ranked search page; array order is the ranking. Ranked search is not
        cursorable in the existing service, so this page intentionally has no cursor field.
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, request_id, details]
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        request_id:
          type: string
          minLength: 1
        details:
          type: object
          additionalProperties: true
...